feat(iOS): Add onFileDownload callback to help iOS apps implement file downloads.
`onFileDownload` is called with the URL that you can use to download the file.
When RNCWebView detects that the HTTP response should result in a file download,
`onFileDownload` is called. The client can then provide code to download
the file.

RNCWebView determines that a file download should take place if either of the
following is true:
1. The HTTP response contains a `Content-Disposition` header that is of type
  'attachment'
2. The MIME type of the response cannot be rendered by the iOS WebView

(cherry picked from commit 4d62dd3295145fe5be6c48bdcf09a4005e4700dc)