fix(Android): Fix multiple files upload on Android
Not sure what it was like when first implemented in react-native-community/react-native-webview#60, but currently if user chooses multiple files to upload, only the first one will actually be uploaded.
The reason behind this is that when mulitple files are selected, both `getData` and `getClipData` will not be null, but `getData` only contains the first of those selections.
Reordering file selection logic to consider the multiple file case first.